added new color with gradient when 2players click the same area, pulled out funtion to choose color, corrected types
