+++ packages/client/src/gameContext.ts
@@ -1,10 +1,11 @@
 import React from "react";
-
+import { TypeSymbol } from "./helper/models/Game.models";
+// import { Symbol } from './components/game'
 export interface IGameContextProps {
     isInRoom: boolean;
     setInRoom: (inRoom: boolean) => void;
-    playerSymbol: "x" | "o";
-    setPlayerSymbol: (symbol: "x" | "o") => void;
+    playerSymbol: TypeSymbol;
+    setPlayerSymbol: (symbol: TypeSymbol) => void;
     isPlayerTurn: boolean;
     setPlayerTurn: (turn: boolean) => void;
     isGameStarted: boolean;
@@ -14,7 +15,7 @@ export interface IGameContextProps {
 const defaultState: IGameContextProps = {
     isInRoom: false,
     setInRoom: () => { },
-    playerSymbol: "x",
+    playerSymbol: TypeSymbol.x,
     setPlayerSymbol: () => { },
     isPlayerTurn: false,
     setPlayerTurn: () => { },
